Automatically attempt to load the models for each module. Models can be explicitly not loaded by passing a 'model': False parameter for a given module.
Also add a proxy fixup helper that sets the IP address correctly based
on the X-Real-IP header. I don't use or advise the use of the
X-Forwarded-For header because it can be a list of IPs and I prefer to
not modify headers that have an accepted or standardized meaning.