When a user picks a TTL value from the check boxes, manually specify the interval in a big if/elif block. I haven't observed any strong guarantees from Flask that user input from forms is tamper proof. I'm not worried about SQL Injection so much as I am bogus infinite lifetime data.
Log the session ID of the user logging in. We haven't changed the session ID upon successful login (yet), but will do so in a follow up commit.

Add a SESSION_BYTES config tunable that specifies the number of bytes to use in a session ID. Right now we're using a 200 bit key, or 25 bytes.

Signed-off-by: Sean Chittenden <sean@chittenden.org>